Conditionally `subdir` libtcg
Previously we unconditionally added the `libtcg` subdir and added
sources to `specific_ss` if `CONFIG_LIBTCG` was set. Conditionally
adding the `subdir` is preferable as we don't expose any project
settings if `libtcg` is disabled.

For some reason `CONFIG_LIBTCG` wasn't being set correctly and the
libtcg sources would be excluded from compilation on M1.

Signed-off-by: Anton Johansson <anjo@rev.ng>